requirements.txt
Flask
pdfplumber
pandas
camelot-py[cv]
opencv-python-headless


